% Auto-generated: do not edit by hand
\name{''DashNotice}

\alias{''DashNotice}

\title{DashNotice component}

\description{

}

\usage{
''DashNotice(id=NULL, badge_color=NULL, border_color=NULL, color=NULL,
duration=NULL, icon_color=NULL, max_length=NULL,
notices=NULL, position=NULL, remove_timing=NULL, sort=NULL,
timeout=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{badge_color}{Character. }

\item{border_color}{Character. }

\item{color}{Character. }

\item{duration}{Numeric. }

\item{icon_color}{Character. }

\item{max_length}{Numeric. }

\item{notices}{List of named lists. }

\item{position}{Character. }

\item{remove_timing}{Character. }

\item{sort}{Character. }

\item{timeout}{Character. }

\item{value}{Character | numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

